Threshold signature scheme agnostic with frost feature
